LPD-28681 Move createIndividuals function to individuals.ts file
