+++ modules/test/playwright/tests/osb-faro-web/utils/utils.ts
@@ -5,30 +5,6 @@
 
 import {Page, expect} from '@playwright/test';
 
-import {ApiHelpers} from '../../../helpers/ApiHelpers';
-
-export async function createIndividuals({
-	apiHelpers,
-	names,
-}: {
-	apiHelpers: ApiHelpers;
-	names: string[];
-}) {
-	const individuals = names.map((name) => ({
-		emailAddress: `${name}@liferay.com`,
-		fields: [
-			{dataSourceId: 0, name: 'givenName', value: name},
-			{dataSourceId: 0, name: 'familyName', value: name},
-			{dataSourceId: 0, name: 'email', value: `${name}@liferay.com`},
-		],
-		firstName: name,
-		id: `${name}@liferay.com`,
-		lastName: name,
-	}));
-
-	await apiHelpers.jsonWebServicesOSBAsah.createIndividuals(individuals);
-}
-
 export async function changeTimeFilterTo({
 	cardName,
 	page,
